+++ CHANGELOG.md
@@ -7,7 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.6] - 2022-01-25
+
+### Added
+
+-   Expressions can have defaults
+-   Extra form checking and show warnings
+
+### Changed
+
+-   Secured JWT tokens even more
+
+### Fixed
+
+-   Fixed editable expression bug
+-   Token issue
+
 ## [2.1.5] - 2022-01-20
+
+### Added
+
+-   Start with changelog
+-   Start with version releases
 
 ## [2.1.4] - 2022-01-20
 
@@ -20,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Hide AWX token
 -   Add AWX ca-bundle for certificate verification
 -   JWT tokens, allow multiple devices
+-   Expression field can be `editable`
+
 
 ## [2.1.2] - 2022-01-17
 
@@ -85,6 +108,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Allow change password for current local user
 -   Start tracking versions
 
-[Unreleased]: https://github.com/ansibleguy76/ansibleforms/compare/2.1.5...HEAD
-
+[2.1.6]: https://github.com/ansibleguy76/ansibleforms/compare/2.1.5...2.1.6
 [2.1.5]: https://github.com/ansibleguy76/ansibleforms/compare/2.1.4...2.1.5